requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via tqdm
cryptography==42.0.5
    # via
    #   djangorestframework-simplejwt
    #   pyjwt
defusedxml==0.7.1
    # via python3-openid
distro==1.9.0
    # via openai
dj-rest-auth==6.0.0
    # via -r requirements.in
django==5.0.3
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-allauth==0.61.1
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-rest-passwordreset==1.4.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
djangorestframework-oauth==1.1.0
    # via -r requirements.in
djangorestframework-simplejwt[crypto]==5.3.1
    # via -r requirements.in
exceptiongroup==1.2.0
    # via anyio
google-search-results==2.4.2
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via openai
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
joblib==1.3.2
    # via scikit-learn
mercadopago==2.2.3
    # via -r requirements.in
numpy==1.26.4
    # via
    #   scikit-learn
    #   scipy
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.14.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic==2.6.4
    # via openai
pydantic-core==2.16.3
    # via pydantic
pyjwt[crypto]==2.8.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
python-dotenv==1.0.1
    # via -r requirements.in
python3-openid==3.2.0
    # via django-allauth
pytz==2024.1
    # via djangorestframework
requests==2.31.0
    # via
    #   django-allauth
    #   google-search-results
    #   mercadopago
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
scikit-learn==1.3.2
    # via -r requirements.in
scipy==1.12.0
    # via scikit-learn
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlparse==0.4.4
    # via django
threadpoolctl==3.3.0
    # via scikit-learn
tqdm==4.66.2
    # via openai
typing-extensions==4.10.0
    # via
    #   anyio
    #   asgiref
    #   openai
    #   pydantic
    #   pydantic-core
tzdata==2024.2
    # via django
urllib3==2.2.1
    # via requests
whitenoise==6.7.0
    # via -r requirements.in
